Move CommitmentConfig into commitment module
